# open-api-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Task Management API",
    "version": "1.0.0",
    "description": "API for managing tasks"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local server"
    }
  ],
  "paths": {
    "/task": {
      "post": {
        "operationId": "createTask",
        "summary": "Create a new task",
        "description": "Creates a new task with the given details",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "My Task"
                  },
                  "content": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": ["Subtask detail 1", "Subtask detail 2"]
                  },
                  "completed": {
                    "type": "boolean",
                    "example": false
                  },
                  "startAt": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2023-08-01T00:00:00Z"
                  },
                  "finishAt": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2023-08-02T00:00:00Z"
                  },
                  "daily": {
                    "type": "boolean",
                    "example": false
                  }
                },
                "required": ["title", "content"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "example": 1
                    },
                    "title": {
                      "type": "string",
                      "example": "My Task"
                    },
                    "content": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "example": ["Task detail 1", "Task detail 2"]
                    },
                    "completed": {
                      "type": "boolean",
                      "example": false
                    },
                    "startAt": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2023-08-01T00:00:00Z"
                    },
                    "finishAt": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2023-08-02T00:00:00Z"
                    },
                    "daily": {
                      "type": "boolean",
                      "example": false
                    },
                    "userId": {
                      "type": "integer",
                      "example": 1
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "example": ["Title is required", "Content is required"]
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "x-channel",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "558881858740@s.whatsapp.net"
          }
        ]
      },
      "get": {
        "operationId": "getTasks",
        "summary": "Get all tasks",
        "description": "Returns a list of all tasks",
        "responses": {
          "200": {
            "description": "A list of tasks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "example": 1
                      },
                      "title": {
                        "type": "string",
                        "example": "My Task"
                      },
                      "content": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        },
                        "example": ["Task detail 1", "Task detail 2"]
                      },
                      "completed": {
                        "type": "boolean",
                        "example": false
                      },
                      "startAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-08-01T00:00:00Z"
                      },
                      "finishAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-08-02T00:00:00Z"
                      },
                      "daily": {
                        "type": "boolean",
                        "example": false
                      },
                      "userId": {
                        "type": "integer",
                        "example": 1
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "errors": {
                      "type": "string",
                      "example": "User not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "x-channel",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "558881858740@s.whatsapp.net"
          }
        ]
      }
    },
    "/task/{id}": {
      "put": {
        "operationId": "updateTask",
        "summary": "Update a task",
        "description": "Updates the details of an existing task",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "Updated Task"
                  },
                  "content": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": ["Updated detail 1", "Updated detail 2"]
                  },
                  "completed": {
                    "type": "boolean",
                    "example": true
                  },
                  "startAt": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2023-08-01T00:00:00Z"
                  },
                  "finishAt": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2023-08-02T00:00:00Z"
                  },
                  "daily": {
                    "type": "boolean",
                    "example": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "example": 1
                    },
                    "title": {
                      "type": "string",
                      "example": "Updated Task"
                    },
                    "content": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "example": ["Updated detail 1", "Updated detail 2"]
                    },
                    "completed": {
                      "type": "boolean",
                      "example": true
                    },
                    "startAt": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2023-08-01T00:00:00Z"
                    },
                    "finishAt": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2023-08-02T00:00:00Z"
                    },
                    "daily": {
                      "type": "boolean",
                      "example": false
                    },
                    "userId": {
                      "type": "integer",
                      "example": 1
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "example": ["Invalid task ID"]
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "x-channel",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "558881858740@s.whatsapp.net"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "example": 1
          }
        ]
      },
      "delete": {
        "operationId": "deleteTask",
        "summary": "Delete a task",
        "description": "Deletes an existing task",
        "responses": {
          "200": {
            "description": "Task deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "example": 1
                    },
                    "title": {
                      "type": "string",
                      "example": "Deleted Task"
                    },
                    "content": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "example": ["Deleted detail 1", "Deleted detail 2"]
                    },
                    "startAt": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2023-08-01T00:00:00Z"
                    },
                    "userId": {
                      "type": "integer",
                      "example": 1
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "errors": {
                      "type": "string",
                      "example": "Invalid task ID"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "x-channel",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "558881858740@s.whatsapp.net"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "example": 1
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "Task": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "title": {
            "type": "string",
            "example": "My Task"
          },
          "content": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["Task detail 1", "Task detail 2"]
          },
          "completed": {
            "type": "boolean",
            "example": false
          },
          "startAt": {
            "type": "string",
            "format": "date-time",
            "example": "2023-08-01T00:00:00Z"
          },
          "finishAt": {
            "type": "string",
            "format": "date-time",
            "example": "2023-08-02T00:00:00Z"
          },
          "daily": {
            "type": "boolean",
            "example": false
          },
          "userId": {
            "type": "integer",
            "example": 1
          }
        }
      }
    }
  }
}